+++ 1.16_combat-6/src/main/java/io/github/axolotlclient/AxolotlClient.java
@@ -22,7 +22,6 @@
 
 package io.github.axolotlclient;
 
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -69,12 +68,12 @@ public class AxolotlClient implements ClientModInitializer {
 
 	public static final String MODID = "axolotlclient";
 	public static final HashMap<Identifier, Resource> runtimeResources = new HashMap<>();
-	public static final Identifier badgeIcon = new Identifier("axolotlclient", "textures/badge.png");
+	public static final Identifier badgeIcon = new Identifier(MODID, "textures/badge.png");
 	public static final OptionCategory config = OptionCategory.create("storedOptions");
 	public static final BooleanOption someNiceBackground = new BooleanOption("defNoSecret", false);
 	public static final List<Module> modules = new ArrayList<>();
+	public static final Logger LOGGER = new LoggerImpl();
 	public static String VERSION;
-	public static Logger LOGGER;
 	public static AxolotlClientConfig CONFIG;
 	public static ConfigManager configManager;
 
@@ -101,14 +100,8 @@ public class AxolotlClient implements ClientModInitializer {
 		modules.addAll(ModuleLoader.loadExternalModules());
 	}
 
-	public static void tickClient() {
-		modules.forEach(Module::tick);
-	}
-
 	@Override
 	public void onInitializeClient() {
-
-		LOGGER = new LoggerImpl();
 
 		VERSION = FabricLoader.getInstance().getModContainer(MODID).orElseThrow(IllegalStateException::new)
 			.getMetadata().getVersion().getFriendlyString();
@@ -118,7 +111,6 @@ public class AxolotlClient implements ClientModInitializer {
 
 		getModules();
 		addExternalModules();
-
 		CONFIG.init();
 
 		new AxolotlClientCommon(LOGGER, Notifications.getInstance(), () -> configManager);
@@ -130,7 +122,7 @@ public class AxolotlClient implements ClientModInitializer {
 		CONFIG.config.add(config);
 
 		io.github.axolotlclient.AxolotlClientConfig.api.AxolotlClientConfig.getInstance()
-			.register(configManager = new VersionedJsonConfigManager(FabricLoader.getInstance().getConfigDir().resolve("AxolotlClient.json"),
+			.register(configManager = new VersionedJsonConfigManager(AxolotlClientCommon.getInstance().getMainConfigFile(),
 				CONFIG.config, 2, (oldVersion, newVersion, config, json) -> {
 				if (oldVersion.getMajor() == 1) {
 					var keystrokes = json.get("hud").getAsJsonObject().get("keystrokehud")
@@ -150,16 +142,12 @@ public class AxolotlClient implements ClientModInitializer {
 
 		modules.forEach(Module::lateInit);
 
-		ClientTickEvents.END_CLIENT_TICK.register(client -> tickClient());
+		ClientTickEvents.END_CLIENT_TICK.register(client -> modules.forEach(Module::tick));
 
 		FeatureDisabler.init();
 
 		LOGGER.debug("Debug Output activated, Logs will be more verbose!");
 
 		LOGGER.info("AxolotlClient Initialized");
-	}
-
-	public static Path resolveConfigFile(String file) {
-		return AxolotlClientCommon.resolveConfigFile(file);
 	}
 }