+++ 1.20/src/main/java/io/github/axolotlclient/AxolotlClient.java
@@ -22,7 +22,6 @@
 
 package io.github.axolotlclient;
 
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -72,7 +71,7 @@ public class AxolotlClient implements ClientModInitializer {
 
 	public static final String MODID = "axolotlclient";
 	public static final HashMap<Identifier, Resource> runtimeResources = new HashMap<>();
-	public static final Identifier badgeIcon = new Identifier("axolotlclient", "textures/badge.png");
+	public static final Identifier badgeIcon = new Identifier(MODID, "textures/badge.png");
 	public static final OptionCategory config = OptionCategory.create("storedOptions");
 	public static final BooleanOption someNiceBackground = new BooleanOption("defNoSecret", false);
 	public static final List<Module> modules = new ArrayList<>();
@@ -105,10 +104,6 @@ public class AxolotlClient implements ClientModInitializer {
 		modules.addAll(ModuleLoader.loadExternalModules());
 	}
 
-	public static void tickClient() {
-		modules.forEach(Module::tick);
-	}
-
 	@Override
 	public void onInitializeClient() {
 
@@ -119,7 +114,6 @@ public class AxolotlClient implements ClientModInitializer {
 
 		getModules();
 		addExternalModules();
-
 		CONFIG.init();
 
 		new AxolotlClientCommon(LOGGER, Notifications.getInstance(), () -> configManager);
@@ -131,7 +125,7 @@ public class AxolotlClient implements ClientModInitializer {
 		CONFIG.getConfig().add(config);
 
 		io.github.axolotlclient.AxolotlClientConfig.api.AxolotlClientConfig.getInstance()
-			.register(configManager = new VersionedJsonConfigManager(FabricLoader.getInstance().getConfigDir().resolve("AxolotlClient.json"),
+			.register(configManager = new VersionedJsonConfigManager(AxolotlClientCommon.getInstance().getMainConfigFile(),
 				CONFIG.getConfig(), 2, (oldVersion, newVersion, config, json) -> {
 				if (oldVersion.getMajor() == 1) {
 					var keystrokes = json.get("hud").getAsJsonObject().get("keystrokehud")
@@ -151,7 +145,7 @@ public class AxolotlClient implements ClientModInitializer {
 
 		modules.forEach(Module::lateInit);
 
-		ClientTickEvents.END_CLIENT_TICK.register(client -> tickClient());
+		ClientTickEvents.END_CLIENT_TICK.register(client -> modules.forEach(Module::tick));
 		ResourceManagerHelper.get(ResourceType.CLIENT_RESOURCES).registerReloadListener(SkyResourceManager.getInstance());
 
 		FeatureDisabler.init();
@@ -159,9 +153,5 @@ public class AxolotlClient implements ClientModInitializer {
 		LOGGER.debug("Debug Output activated, Logs will be more verbose!");
 
 		LOGGER.info("AxolotlClient Initialized");
-	}
-
-	public static Path resolveConfigFile(String file) {
-		return AxolotlClientCommon.resolveConfigFile(file);
 	}
 }