+++ 1.21/src/main/java/io/github/axolotlclient/modules/hud/gui/keystrokes/KeyBindsList.java
@@ -108,7 +108,7 @@ public class KeyBindsList extends ElementListWidget<KeyBindsList.Entry> {
 	public class KeyEntry extends Entry {
 		private static final Text REMOVE_BUTTON_TITLE = Text.translatable("keystrokes.stroke.remove");
 		private final KeystrokeHud.Keystroke key;
-		private Text name;
+		private final Text name;
 		private final ButtonWidget configureButton, removeButton;
 
 		KeyEntry(final KeystrokeHud.Keystroke key, final Text name) {
@@ -120,6 +120,7 @@ public class KeyBindsList extends ElementListWidget<KeyBindsList.Entry> {
 			this.removeButton = ButtonWidget.builder(REMOVE_BUTTON_TITLE, b -> {
 					removeEntry(this);
 					keyBindsScreen.removeKey(key);
+					refreshScrollAmount();
 				}).positionAndSize(0, 0, 50, 20)
 				.build();
 		}