+++ 1.8.9/src/main/java/io/github/axolotlclient/AxolotlClient.java
@@ -22,7 +22,6 @@
 
 package io.github.axolotlclient;
 
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -69,7 +68,7 @@ public class AxolotlClient implements ClientModInitializer {
 
 	public static final String MODID = "axolotlclient";
 	public static final HashMap<Identifier, Resource> runtimeResources = new HashMap<>();
-	public static final Identifier badgeIcon = new Identifier("axolotlclient", "textures/badge.png");
+	public static final Identifier badgeIcon = new Identifier(MODID, "textures/badge.png");
 	public static final OptionCategory config = OptionCategory.create("storedOptions");
 	public static final BooleanOption someNiceBackground = new BooleanOption("defNoSecret", false);
 	public static final List<Module> modules = new ArrayList<>();
@@ -102,10 +101,6 @@ public class AxolotlClient implements ClientModInitializer {
 		modules.addAll(ModuleLoader.loadExternalModules());
 	}
 
-	public static void tickClient() {
-		modules.forEach(Module::tick);
-	}
-
 	@Override
 	public void onInitializeClient() {
 
@@ -127,7 +122,7 @@ public class AxolotlClient implements ClientModInitializer {
 
 		CONFIG.getConfig().add(config);
 
-		AxolotlClientConfig.getInstance().register(configManager = new VersionedJsonConfigManager(FabricLoader.getInstance().getConfigDir().resolve("AxolotlClient.json"),
+		AxolotlClientConfig.getInstance().register(configManager = new VersionedJsonConfigManager(AxolotlClientCommon.getInstance().getMainConfigFile(),
 			CONFIG.getConfig(), 2, (oldVersion, newVersion, config, json) -> {
 			if (oldVersion.getMajor() == 1) {
 				var keystrokes = json.get("hud").getAsJsonObject().get("keystrokehud")
@@ -147,16 +142,12 @@ public class AxolotlClient implements ClientModInitializer {
 
 		modules.forEach(Module::lateInit);
 
-		MinecraftClientEvents.TICK_END.register(client -> tickClient());
+		MinecraftClientEvents.TICK_END.register(client -> modules.forEach(Module::tick));
 
 		FeatureDisabler.init();
 
 		LOGGER.debug("Debug Output enabled, Logs will be quite verbose!");
 
 		LOGGER.info("AxolotlClient Initialized");
-	}
-
-	public static Path resolveConfigFile(String file) {
-		return AxolotlClientCommon.resolveConfigFile(file);
 	}
 }